Add DELETE endpoint to AuctionsController
Added a new HTTP DELETE endpoint `DeleteAuction` to the `AuctionsController` class in the `AuctionService.Controllers` namespace. This endpoint allows for the deletion of an auction by its `id`. If the auction is not found, a `NotFound` response is returned. If the auction is found, it is removed from the database, changes are saved, and an `Ok` response is returned.